+++ structurizr-dsl/src/main/java/com/structurizr/dsl/StructurizrDslTokens.java
@@ -60,6 +60,8 @@ class StructurizrDslTokens {
     static final String KROKI_TOKEN = "kroki";
     static final String IMAGE_TOKEN = "image";
     static final String STYLES_TOKEN = "styles";
+    static final String LIGHT_COLOR_SCHEME_TOKEN = "light";
+    static final String DARK_COLOR_SCHEME_TOKEN = "dark";
     static final String BRANDING_TOKEN = "branding";
     static final String BRANDING_LOGO_TOKEN = "logo";
     static final String BRANDING_FONT_TOKEN = "font";