+++ structurizr-dsl/src/main/java/com/structurizr/dsl/StylesDslContext.java
@@ -1,6 +1,22 @@
 package com.structurizr.dsl;
 
+import com.structurizr.view.ColorScheme;
+
 final class StylesDslContext extends DslContext {
+
+    private final ColorScheme colorScheme;
+
+    StylesDslContext() {
+        colorScheme = null;
+    }
+
+    StylesDslContext(ColorScheme colorScheme) {
+        this.colorScheme = colorScheme;
+    }
+
+    ColorScheme getColorScheme() {
+        return colorScheme;
+    }
 
     @Override
     protected String[] getPermittedTokens() {