+++ structurizr-dsl/src/test/java/com/structurizr/dsl/ElementStyleParserTests.java
@@ -23,10 +23,17 @@ class ElementStyleParserTests extends AbstractTests {
         return context;
     }
 
+    private StylesDslContext stylesDslContext() {
+        StylesDslContext context = new StylesDslContext();
+        context.setWorkspace(workspace);
+
+        return context;
+    }
+
     @Test
     void test_parseElementStyle_ThrowsAnException_WhenThereAreTooManyTokens() {
         try {
-            parser.parseElementStyle(context(), tokens("element", "tag", "extra"));
+            parser.parseElementStyle(stylesDslContext(), tokens("element", "tag", "extra"));
             fail();
         } catch (Exception e) {
             assertEquals("Too many tokens, expected: element <tag> {", e.getMessage());
@@ -36,7 +43,7 @@ class ElementStyleParserTests extends AbstractTests {
     @Test
     void test_parseElementStyle_ThrowsAnException_WhenTheTagIsMissing() {
         try {
-            parser.parseElementStyle(context(), tokens("element"));
+            parser.parseElementStyle(stylesDslContext(), tokens("element"));
             fail();
         } catch (Exception e) {
             assertEquals("Expected: element <tag> {", e.getMessage());
@@ -46,7 +53,7 @@ class ElementStyleParserTests extends AbstractTests {
     @Test
     void test_parseElementStyle_ThrowsAnException_WhenTheTagIsEmpty() {
         try {
-            parser.parseElementStyle(context(), tokens("element", ""));
+            parser.parseElementStyle(stylesDslContext(), tokens("element", ""));
             fail();
         } catch (Exception e) {
             assertEquals("A tag must be specified", e.getMessage());
@@ -55,7 +62,7 @@ class ElementStyleParserTests extends AbstractTests {
 
     @Test
     void test_parseElementStyle_CreatesAnElementStyle() {
-        parser.parseElementStyle(context(), tokens("element", "Element"));
+        parser.parseElementStyle(stylesDslContext(), tokens("element", "Element"));
 
         ElementStyle style = workspace.getViews().getConfiguration().getStyles().getElements().stream().filter(es -> "Element".equals(es.getTag())).findFirst().get();
         assertNotNull(style);
@@ -64,7 +71,7 @@ class ElementStyleParserTests extends AbstractTests {
     @Test
     void test_parseElementStyle_FindsAnExistingElementStyle() {
         ElementStyle style = workspace.getViews().getConfiguration().getStyles().addElementStyle("Tag");
-        assertSame(style, parser.parseElementStyle(context(), tokens("element", "Tag")));
+        assertSame(style, parser.parseElementStyle(stylesDslContext(), tokens("element", "Tag")));
     }
 
     @Test